Move component TokyoRulesApplicationNumberCard from 'components/cards' to 'components/index/CardsMonitoring'
